fix issue #299: arithmetic operations in brackets are comma separated
for example the query:

```sql
SELECT (1 + 2) AS THREE, (1 + 3) AS FOUR
```

would become:

```sql
SELECT (1,+,2) AS THREE, (1 + 3) AS FOUR
```

This is a regression caused by commit 3a9d906, that meant to add missing
commas in expressions like `CAST(... AS DECIMAL(16, 2))`,
but added too many commas visibly.

=> revert the parts of that commit that changed the code (leave the
added test assertion)

=> change the `ExpressionListProcessor` code that handles
unspecified nodes.

That code was added in commit 63d59d7, apparently to solve issue 51:
https://code.google.com/archive/p/php-sql-parser/issues/51

It seems to have copied and pasted the code that handles function
arguments, just above.

But that code does not keep commas,
and while the `FunctionBuilder` adds them back,
the `SelectBracketExpressionBuilder` does not.

The issue51Test test seems to cover the bug described in issue 51,
and it still passes.

So maybe we do not need the same handling as for function arguments in
unspecified nodes?

Hopefully someone with a better understanding of this code can have a look.
